Finished rewriting TCD.InteropServices.
