+++ sources/TCD.ComponentModel/Assembly.cs
@@ -1,0 +1,9 @@
+/***************************************************************************************************
+ * FileName:             Assembly.cs
+ * Copyright:             Copyright © 2017-2019 Thomas Corwin, et al. All Rights Reserved.
+ * License:              https://github.com/tom-corwin/tcdfx/blob/master/LICENSE.md
+ **************************************************************************************************/
+
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("TCD.InteropServices")]
